Make sure to not match on every zk option.
The script allows for the passing of zk run commands as files under the
proper conf_dir. If an option like zk_timeout is changed with this
method, the wildcard will attempt to match that as the option to pass to
the --zk command. The extra space after the --zk line in this script
should only match if someone is actually passing that exact command.